plugins/shellz/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
pathlib==1.0.1
pycparser==2.21
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
websocket-client==1.5.1
websockets==10.4
zope.event==4.6
zope.interface==6.0